add CRUD operation in CinemaService, and change data type in all Jpa Repositories

+++ src/main/java/com/example/bookingsystem/repository/CinemaRepository.java
@@ -4,5 +4,5 @@ import com.example.bookingsystem.entity.CinemaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface CinemaRepository extends JpaRepository<CinemaEntity, Integer>, JpaSpecificationExecutor<CinemaEntity> {
+public interface CinemaRepository extends JpaRepository<CinemaEntity, Long>, JpaSpecificationExecutor<CinemaEntity> {
 }
