+++ src/main/java/com/example/bookingsystem/repository/RentRepository.java
@@ -4,5 +4,5 @@ import com.example.bookingsystem.entity.RentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface RentRepository extends JpaRepository<RentEntity, Integer>, JpaSpecificationExecutor<RentEntity> {
+public interface RentRepository extends JpaRepository<RentEntity, Long>, JpaSpecificationExecutor<RentEntity> {
 }
